Greener tracks
Shift the hue of the tracks from 214° (very close to the hue of mixed-cycle-fill, which is 217°) to 140° (matching the hue of nomotor-fill), while maintaining the lightness and saturation. For information, the hue of path-fill is 170° and cycle-fill is 240°, so with this change there is a continuum from roads that are (mostly) closed to motor vehicles (greenest) through paths that are either allowed or designated for bikes, but not reserved, to dedicated cycleways (bluest).