Agricultural/forestry roads should not be considered bikable
As reported by a user:

> It has come to my attention that for Austria CyclOSM is not rendering
roads marked as vehicle=forestry correctly. According to Austrian law,
forestry roads are not allowed for bicycles, unless specifically
dedicated for them by signage. CyclOSM is not making a difference
though, between roads tagged as vehicle=forestry and roads not tagged
this access-tag.

See
https://discuss.graphhopper.com/t/bicycle-allowed-on-vehicle-delivery-or-vehicle-forestry/6983/3
as well.